// src/utils/enums.ts
export enum statuses {
    PROCESSED = 'PROCESSED',
    ACCEPTED = 'ACCEPTED',
    DECLINE = 'DECLINE'
}

export enum queryTypes {
    CHANGE_ROLE = 'CHANGE_ROLE',
    JOIN_TEAM = 'JOIN_TEAM',
    EXIT_FROM_TEAM = 'EXIT_FROM_TEAM',
    MOVE_TO_ANOTHER_TEAM = 'MOVE_TO_ANOTHER_TEAM'
}

export enum userRoles {
    PLAYER = "PLAYER",
    MANAGER = "MANAGER",
    ADMIN = "ADMIN"
}